Add connectionCleared method
This adds a .connectionCleared method, which is called on an object when it's connection is cleared. This allows us to do clean-up in cases where we know our connection is going away. Or - more specifically - it allows us to AVOID doing cleanup (if desired) in cases where a disconnect is only temporary.