Refactor UpdateDispatcher
This makes UpdateDispatchers automatically clean up. How? We use a Connection privately to track the link between out UpdateDispatcher and the UpdateForwarder. We return THIS CONNECTION - so any connection that's chained off of it will notify it when it's cleared. The connection calls us back via connectionCleared, which we can use to tear down the related objects.