Send both input (unmapped) and value (mapped) when notifying ControlValue dependents
Both the unmapped and mapped value are important depending on what you're connecting a ControlValue to. Getting the unmapped value (the input) is relatively verbose considering how common a use case it is, so this provides two ways to get the unmapped value: it is passed as an extra argument to value notification, and it's available by registering for the \input signal. This makes, for example, mapping a ControlValue to a synth AND to a slider as simple as:
   cv.signal(\value).connectTo(synth.argSlot(\freq));
   cv.signal(\input).connectTo(slider.valueSlot);