Fix value forwarding on initial connection
Make UpdateFilter and UpdateTransform subclass UpdateForwarderWithInit,
which keeps track of last object sending a
\changed message. Everytime a dependant is added to the forwarder,
this object is asked to resend its value.

This is a bit of a kludge, but fixes forwarding initial values through a
chain of several forwarders, which would otherwise stop at first
UpdateForwarder subclass when the value is sent before the dependency
chain is ready.

The value will be sent once for every
link in the forwarder chain, which might build up and cause issues for
complex chains.